refactor: move split equalizer to own file
